requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Table parsing
pdfplumber>=0.9.0
tabula-py>=2.5.0
camelot-py[cv]>=0.11.0

# Knowledge Graph
rdflib>=6.3.0
networkx>=3.1
pyvis>=0.3.1

# RAG and embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
langchain>=0.1.0
openai>=1.0.0

# Korean NLP
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Evaluation
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.0
scipy>=1.10.0
zss>=1.2.0
ragas>=0.1.0
datasets>=2.14.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Web scraping (for data collection)
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0

# Baseline models (optional, install as needed)
huggingface_hub>=0.16.0

